backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "backend for the course"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==25.1.0",
    "aiohttp==3.13.0",
    "alembic==1.17.0",
    "asyncpg==0.30.0",
    "black==25.9.0",
    "boto3==1.40.53",
    "celery==5.5.3",
    "elasticsearch==9.1.1",
    "email-validator==2.3.0",
    "fastapi==0.119",
    "ffmpeg-python==0.2.0",
    "flower==2.0.1",
    "gunicorn==23.0.0",
    "httpx==0.28.1",
    "jinja2==3.1.6",
    "bcrypt==5.0.0",
    "pillow==12.0.0",
    "prometheus-client==0.23.1",
    "psycopg2-binary==2.9.11",
    "pydantic==2.12.2",
    "pydantic-settings==2.11.0",
    "pypdf2==3.0.1",
    "pytest==8.4.2",
    "pytest-asyncio==1.2.0",
    "pytest-cov==7.0.0",
    "python-dotenv==1.1.1",
    "python-jose[cryptography]==3.5.0",
    "python-json-logger==4.0.0",
    "python-multipart==0.0.20",
    "python-pptx==1.0.2",
    "redis==6.4.0",
    "ruff==0.14.0",
    "sentry-sdk[fastapi]==2.42.0",
    "sqlalchemy==2.0.44",
    "structlog==25.4.0",
    "uvicorn[standard]==0.37",
]
